fix(cli): Ensure aspect cache dir is a directory (#5980)
Creates a method `ioutils.AspectCacheDir` which
checks that the cache directory is a valid dir.
If it is not, it deletes whatever exists at that
path and creates a fresh dir

Fixes #1226

---

### Changes are visible to end-users: no

GitOrigin-RevId: efe384e523c232586df938c00a37ef098ceebe9a
