+++ pkg/ioutils/BUILD.bazel
@@ -3,10 +3,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "ioutils",
     srcs = [
+        "cache.go",
         "prompt.go",
         "streams.go",
     ],
     importpath = "aspect.build/cli/pkg/ioutils",
     visibility = ["//visibility:public"],
-    deps = ["@com_github_manifoldco_promptui//:promptui"],
+    deps = [
+        "@com_github_manifoldco_promptui//:promptui",
+        "@com_github_mitchellh_go_homedir//:go-homedir",
+    ],
 )
